feat: Add new columns 'framework' and 'paymentMethod' to Task table
This commit adds two new columns, 'framework' and 'paymentMethod', to the Task table in the database. The migration script creates a new table called "new_Task" with the new columns, copies the data from the existing "Task" table, and then renames the new table to "Task". This change allows for storing additional information related to tasks.